Factorize copy paste
